pytorch_laia/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pillow>=9.0.0
tqdm>=4.65.0
tensorboard>=2.13.0
einops>=0.6.0
pytorch-lightning>=2.0.0  # For structured training loops
wandb>=0.15.0  # For experiment tracking 